requirements.txt
# SEFNet: Selective Equivariant Features for Robust Scale-Adaptive UAV Tracking
# Tested with Python 3.8+ and CUDA 11.8

# Core
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0                # ViT-Base pretrained backbone

# Scientific computing
numpy>=1.22.0
scipy>=1.9.0               # Gaussian derivatives, special functions
einops>=0.6.0              # Tensor rearrangement for attention

# Training
tensorboard>=2.12.0
wandb>=0.15.0              # Optional: experiment tracking
pyyaml>=6.0
easydict>=1.10

# Data & evaluation
opencv-python>=4.7.0
Pillow>=9.4.0
pycocotools>=2.0.6
lmdb>=1.4.0               # Optional: fast data loading
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0            # Feature decomposition plots (Fig.7)

# Misc
yacs>=0.1.8
colorama>=0.4.6
